[K/N] Move `enumEntriesMap` helper into the common backend
We are going to reuse it in Kotlin/JS for KT-63073.
