+++ compiler/ir/backend.common/src/org/jetbrains/kotlin/backend/common/lower/EnumWhenLowering.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2026 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
@@ -10,6 +10,7 @@ import org.jetbrains.kotlin.backend.common.FileLoweringPass
 import org.jetbrains.kotlin.backend.common.IrElementTransformerVoidWithContext
 import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.ir.builders.*
+import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrEnumEntry
 import org.jetbrains.kotlin.ir.declarations.IrFile
 import org.jetbrains.kotlin.ir.declarations.IrVariable
@@ -17,9 +18,12 @@ import org.jetbrains.kotlin.ir.expressions.*
 import org.jetbrains.kotlin.ir.expressions.impl.IrCallImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrConstImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrGetValueImpl
+import org.jetbrains.kotlin.ir.irAttribute
 import org.jetbrains.kotlin.ir.types.classifierOrNull
 import org.jetbrains.kotlin.ir.types.getClass
 import org.jetbrains.kotlin.ir.util.*
+import org.jetbrains.kotlin.name.Name
+import org.jetbrains.kotlin.utils.addToStdlib.getOrSetIfNull
 
 /**
  * Optimization: replaces `when` subjects of enum types with their ordinals.
@@ -170,4 +174,22 @@ open class EnumWhenLowering(protected open val context: CommonBackendContext) : 
         }
     }
 
+}
+
+private var IrClass.enumEntriesMap: Map<Name, LoweredEnumEntryDescription>? by irAttribute(copyByDefault = false)
+
+data class LoweredEnumEntryDescription(val ordinal: Int, val getterId: Int)
+
+fun enumEntriesMap(enumClass: IrClass): Map<Name, LoweredEnumEntryDescription> {
+    require(enumClass.isEnumClass) { "Expected enum class but was: ${enumClass.render()}" }
+    return enumClass::enumEntriesMap.getOrSetIfNull {
+        data class NameWithOrdinal(val name: Name, val ordinal: Int)
+        enumClass.declarations.asSequence()
+            .filterIsInstance<IrEnumEntry>()
+            .mapIndexed { index, it -> NameWithOrdinal(it.name, index) }
+            .sortedBy { it.name }
+            .withIndex()
+            .associate { it.value.name to LoweredEnumEntryDescription(it.value.ordinal, it.index) }
+            .toMap()
+    }
 }
